{"schema":"decodex/commit/1","summary":"Modularize retry and harness runtime surfaces","authority":"manual"}
