+++ apps/decodex/src/orchestrator/harness_improvement/model_constants.rs
@@ -1,0 +1,2 @@
+pub(super) const HARNESS_OUTCOME_SCHEMA: &str = "decodex.harness_outcome/1";
+pub(super) const HARNESS_OUTCOME_EVENT_TYPE: &str = "harness_outcome";
