+++ apps/decodex/src/orchestrator/status/post_review/retry_budget.rs
@@ -1,205 +1,14 @@
-use crate::orchestrator::status::post_review::{
-	self, Command, Path, PostReviewLaneClassification, PostReviewLaneDecision,
-	PostReviewLaneSnapshot, PullRequestMergeViewResponse, PullRequestReadbackRootCause,
-	PullRequestReviewState, PullRequestReviewStateInspector, ReviewHandoffMarker, ServiceConfig,
-	WorkflowDocument, github,
+mod closeout_confirmation;
+mod exhausted;
+mod review_state;
+mod worktree_status;
+
+pub(crate) use self::{
+	closeout_confirmation::confirm_status_visible_merged_closeout,
+	exhausted::{
+		merged_closeout_pending_classification,
+		retry_budget_exhausted_post_review_lane_classification,
+	},
+	review_state::retry_budget_exhausted_merged_review_state,
+	worktree_status::worktree_has_no_tracked_changes,
 };
-
-pub(crate) fn retry_budget_exhausted_post_review_lane_classification<I>(
-	snapshot: &PostReviewLaneSnapshot,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	review_state_inspector: &I,
-	mut classification: PostReviewLaneClassification,
-) -> PostReviewLaneClassification
-where
-	I: PullRequestReviewStateInspector,
-{
-	if classification.pr_url.is_none() {
-		classification.pr_url =
-			snapshot.review_handoff.as_ref().map(|marker| marker.pr_url().to_owned());
-	}
-	if classification.pr_state.is_none()
-		&& let Some(review_state) =
-			retry_budget_exhausted_merged_review_state(snapshot, review_state_inspector)
-	{
-		classification = post_review::initial_post_review_lane_classification(&review_state);
-
-		post_review::apply_pre_orchestration_post_review_classification(
-			snapshot,
-			workflow,
-			&review_state,
-			&mut classification,
-		);
-	}
-	if merged_closeout_pending_classification(&classification)
-		&& worktree_has_no_tracked_changes(snapshot.worktree.worktree_path())
-	{
-		if snapshot.issue.state.name == workflow.frontmatter().tracker().resolved_completed_state()
-			&& !worktree_has_no_tracked_changes(project.repo_root())
-		{
-			classification.decision = PostReviewLaneDecision::CleanupBlocked;
-			classification.reason = String::from("default_branch_worktree_dirty");
-
-			return post_review::finalize_post_review_lane_classification_with_retry_budget(
-				snapshot,
-				classification,
-				true,
-			);
-		}
-
-		return post_review::finalize_post_review_lane_classification(snapshot, classification);
-	}
-	if classification.pr_state.as_deref() == Some("MERGED")
-		&& worktree_has_no_tracked_changes(snapshot.worktree.worktree_path())
-	{
-		classification.decision = if snapshot.issue.state.name
-			== workflow.frontmatter().tracker().resolved_completed_state()
-		{
-			PostReviewLaneDecision::CleanupBlocked
-		} else {
-			PostReviewLaneDecision::CloseoutBlocked
-		};
-		classification.reason = String::from("retry_budget_exhausted");
-
-		return post_review::finalize_post_review_lane_classification_with_retry_budget(
-			snapshot,
-			classification,
-			true,
-		);
-	}
-
-	classification.decision = PostReviewLaneDecision::Block;
-	classification.reason = String::from("retry_budget_exhausted");
-
-	post_review::finalize_post_review_lane_classification_with_retry_budget(
-		snapshot,
-		classification,
-		true,
-	)
-}
-
-pub(crate) fn merged_closeout_pending_classification(
-	classification: &PostReviewLaneClassification,
-) -> bool {
-	classification.decision == PostReviewLaneDecision::Continue
-		&& classification.reason == "pull_request_merged_closeout_pending"
-		&& classification.pr_state.as_deref() == Some("MERGED")
-}
-
-pub(crate) fn confirm_status_visible_merged_closeout(
-	snapshot: &PostReviewLaneSnapshot,
-	project: &ServiceConfig,
-	classification: &mut PostReviewLaneClassification,
-) {
-	if !merged_closeout_pending_classification(classification) {
-		return;
-	}
-
-	let Some(pr_url) = classification.pr_url.as_deref() else {
-		mark_merged_closeout_confirmation_conflict(classification, None, None);
-
-		return;
-	};
-	let expected_head_sha = snapshot
-		.review_handoff
-		.as_ref()
-		.map(ReviewHandoffMarker::pr_head_oid)
-		.or(classification.pr_head_sha.as_deref());
-	let Some(expected_head_sha) = expected_head_sha else {
-		mark_merged_closeout_confirmation_conflict(classification, None, None);
-
-		return;
-	};
-	let github_token = match post_review::resolve_configured_env_var(
-		"github.token_env_var",
-		Some(project.github().token_env_var()),
-	) {
-		Ok(github_token) => github_token,
-		Err(error) => {
-			let root_cause = post_review::classify_pull_request_readback_report(&error);
-
-			mark_merged_closeout_confirmation_conflict(classification, None, Some(root_cause));
-
-			return;
-		},
-	};
-	let merge_readback = match github::inspect_pull_request_merge_readback(
-		snapshot.worktree.worktree_path(),
-		pr_url,
-		&github_token,
-		project.github().command_path(),
-	) {
-		Ok(merge_readback) => merge_readback,
-		Err(error) => {
-			let root_cause = post_review::classify_pull_request_readback_report(&error);
-
-			mark_merged_closeout_confirmation_conflict(classification, None, Some(root_cause));
-
-			return;
-		},
-	};
-
-	if merge_readback.state == "MERGED"
-		&& merge_readback.head_ref_oid.as_deref() == Some(expected_head_sha)
-	{
-		return;
-	}
-
-	mark_merged_closeout_confirmation_conflict(
-		classification,
-		Some(merge_readback),
-		Some(PullRequestReadbackRootCause::LineageValidationFailed),
-	);
-}
-
-pub(crate) fn mark_merged_closeout_confirmation_conflict(
-	classification: &mut PostReviewLaneClassification,
-	merge_readback: Option<PullRequestMergeViewResponse>,
-	root_cause: Option<PullRequestReadbackRootCause>,
-) {
-	classification.decision = PostReviewLaneDecision::Block;
-	classification.reason = String::from("pull_request_merge_state_conflict");
-	classification.readback_warning = Some(String::from("pull_request_merge_state_conflict"));
-	classification.readback_root_cause =
-		root_cause.map(|root_cause| root_cause.as_str().to_owned());
-
-	if let Some(merge_readback) = merge_readback {
-		classification.pr_state = Some(merge_readback.state);
-		classification.pr_head_sha =
-			merge_readback.head_ref_oid.or_else(|| classification.pr_head_sha.clone());
-	}
-}
-
-pub(crate) fn retry_budget_exhausted_merged_review_state<I>(
-	snapshot: &PostReviewLaneSnapshot,
-	review_state_inspector: &I,
-) -> Option<PullRequestReviewState>
-where
-	I: PullRequestReviewStateInspector,
-{
-	let review_handoff = snapshot.review_handoff.as_ref()?;
-
-	if !worktree_has_no_tracked_changes(snapshot.worktree.worktree_path()) {
-		return None;
-	}
-
-	let review_state = review_state_inspector
-		.inspect_review_state(snapshot.worktree.worktree_path(), review_handoff.pr_url())
-		.ok()?;
-
-	(review_state.state == "MERGED").then_some(review_state)
-}
-
-pub(crate) fn worktree_has_no_tracked_changes(worktree_path: &Path) -> bool {
-	let Ok(output) = Command::new("git")
-		.arg("-C")
-		.arg(worktree_path)
-		.args(["status", "--porcelain", "--untracked-files=no"])
-		.output()
-	else {
-		return false;
-	};
-
-	output.status.success() && String::from_utf8_lossy(&output.stdout).trim().is_empty()
-}
